requirements.txt
python-dotenv~=0.14.0
discord.py
requests~=2.25.1
pymongo~=3.11.4